refactor: reorg ShipCustomizer.lua to Ship/
